@@ -50,8 +50,8 @@
     "@fortawesome/react-fontawesome": "^0.2.0",
     "classnames": "^2.3.1",
     "dayjs": "^1.11.7",
-    "react": "^17.0.2",
-    "react-dom": "17.0.2",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
     "react-is": "^18.0.0",
     "react-transition-group": "^4.4.2"
   },
@@ -62,8 +62,8 @@
     "@types/node": "^17.0.21",
     "@types/react": "^17.0.33",
     "classnames": "^2.3.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
     "react-is": "^18.0.0",
     "typescript": "^4.5.4"
   },
