@@ -1,5 +1,5 @@
 import React from "react";
-import ReactDOM from "react-dom";
+import ReactDOM from "react-dom/client";
 import "./styles/index.scss";
 
 import ExButton from "@/packages/button/__expample__/ExButton";
@@ -7,7 +7,19 @@ import ExInput from "@/packages/input/__expample__/ExInput";
 import EXIcon from "@/packages/icon/__expample__/EXIcon";
 import EXDatePicker from "@/packages/date-picker/__expample__/EXDatePicker";
 
-ReactDOM.render(
+// ReactDOM.render(
+//   <div style={{ gap: 32, display: "flex", flexDirection: "column" }}>
+//     <div>
+//       {/*<ExInput />*/}
+//       {/*<ExButton />*/}
+//       {/*<EXIcon />*/}
+//       <EXDatePicker />
+//     </div>
+//   </div>,
+//   document.getElementById("root")
+// );
+
+ReactDOM.createRoot(document.getElementById("root")!).render(
   <div style={{ gap: 32, display: "flex", flexDirection: "column" }}>
     <div>
       {/*<ExInput />*/}
@@ -15,6 +27,5 @@ ReactDOM.render(
       {/*<EXIcon />*/}
       <EXDatePicker />
     </div>
-  </div>,
-  document.getElementById("root")
+  </div>
 );
